refactor(ai-diagnostic): split drone_integration into models + planner
drone_integration.py had no coverage. Added 9 characterisation tests over the
pure geometry (GeoPoint haversine distance incl. altitude, and the orbit /
tower-spiral / thermal-grid waypoint patterns with counts, radius, distance and
duration), then extracted, test-first:

  service/drone_models.py         - drone/mission/capture enums + GeoPoint,
                                    Waypoint, FlightPath, CapturedData,
                                    DroneState, Mission
  service/flight_path_planner.py  - FlightPathPlanner (pure geometry)

drone_integration.py keeps the DroneController + DroneIntegrationService
orchestration and re-exports every name (importers unchanged). Dropped now-dead
imports (numpy, json, dataclass/field, Enum, Callable, timedelta). 741 -> 320
LOC. All 217 ai-diagnostic tests pass (208 + 9 new).
