+++ ai-diagnostic/service/drone_integration.py
@@ -10,17 +10,12 @@ Provides integration with drone platforms for:
 - Safety zone management
 """
 
-import json
 import logging
 import threading
 import time
 import uuid
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from enum import Enum
-from typing import Any, Callable, Dict, List, Optional, Tuple
-
-import numpy as np
+from datetime import datetime
+from typing import Any, Dict, List, Optional, Tuple
 
 logger = logging.getLogger(__name__)
 
@@ -29,426 +24,15 @@ from .utils.rng import get_rng
 _rng = get_rng()
 
 
-class DroneStatus(Enum):
-    """Drone operational status."""
-    OFFLINE = "offline"
-    STANDBY = "standby"
-    PREFLIGHT = "preflight"
-    FLYING = "flying"
-    HOVERING = "hovering"
-    RETURNING = "returning"
-    LANDING = "landing"
-    CHARGING = "charging"
-    MAINTENANCE = "maintenance"
-    ERROR = "error"
-
-
-class MissionType(Enum):
-    """Types of drone missions."""
-    SITE_INSPECTION = "site_inspection"
-    TOWER_INSPECTION = "tower_inspection"
-    ANTENNA_INSPECTION = "antenna_inspection"
-    THERMAL_SCAN = "thermal_scan"
-    EMERGENCY_RESPONSE = "emergency_response"
-    PERIMETER_CHECK = "perimeter_check"
-    VEGETATION_SURVEY = "vegetation_survey"
-    DAMAGE_ASSESSMENT = "damage_assessment"
-
-
-class MissionStatus(Enum):
-    """Mission execution status."""
-    PLANNED = "planned"
-    QUEUED = "queued"
-    PREFLIGHT_CHECK = "preflight_check"
-    IN_PROGRESS = "in_progress"
-    PAUSED = "paused"
-    RETURNING = "returning"
-    COMPLETED = "completed"
-    ABORTED = "aborted"
-    FAILED = "failed"
-
-
-class CaptureType(Enum):
-    """Types of data capture."""
-    PHOTO = "photo"
-    VIDEO = "video"
-    THERMAL = "thermal"
-    MULTISPECTRAL = "multispectral"
-    LIDAR = "lidar"
-
-
-@dataclass
-class GeoPoint:
-    """Geographic point with altitude."""
-    latitude: float
-    longitude: float
-    altitude: float  # meters above ground level
-    heading: Optional[float] = None  # Heading in degrees (north is 0)
-
-    def to_dict(self) -> Dict[str, Optional[float]]:
-        return {
-            "latitude": self.latitude,
-            "longitude": self.longitude,
-            "altitude": self.altitude,
-            "heading": self.heading
-        }
-
-    def distance_to(self, other: 'GeoPoint') -> float:
-        """Calculate approximate distance in meters."""
-        # Simplified haversine
-        lat1, lon1 = np.radians(self.latitude), np.radians(self.longitude)
-        lat2, lon2 = np.radians(other.latitude), np.radians(other.longitude)
-
-        dlat = lat2 - lat1
-        dlon = lon2 - lon1
-
-        a = np.sin(dlat/2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2)**2
-        c = 2 * np.arcsin(np.sqrt(a))
-
-        r = 6371000  # Earth radius in meters
-        horizontal = r * c
-        vertical = abs(self.altitude - other.altitude)
-
-        return np.sqrt(horizontal**2 + vertical**2)
-
-
-@dataclass
-class Waypoint:
-    """Mission waypoint with actions."""
-    waypoint_id: str
-    position: GeoPoint
-    action: str = "flyover"  # flyover, hover, capture, orbit
-    hover_time: float = 0.0  # seconds
-    capture_type: Optional[CaptureType] = None
-    gimbal_pitch: float = -90.0  # degrees, -90 = straight down
-    speed: float = 5.0  # m/s
-
-    def to_dict(self) -> Dict[str, Any]:
-        return {
-            "waypoint_id": self.waypoint_id,
-            "position": self.position.to_dict(),
-            "action": self.action,
-            "hover_time": self.hover_time,
-            "capture_type": self.capture_type.value if self.capture_type else None,
-            "gimbal_pitch": self.gimbal_pitch,
-            "speed": self.speed
-        }
-
-
-@dataclass
-class FlightPath:
-    """Planned flight path for a mission."""
-    path_id: str
-    waypoints: List[Waypoint]
-    total_distance: float = 0.0  # meters
-    estimated_duration: float = 0.0  # seconds
-    safety_buffer: float = 10.0  # meters from obstacles
-
-    def to_dict(self) -> Dict[str, Any]:
-        return {
-            "path_id": self.path_id,
-            "waypoints": [w.to_dict() for w in self.waypoints],
-            "total_distance": self.total_distance,
-            "estimated_duration": self.estimated_duration,
-            "waypoint_count": len(self.waypoints)
-        }
-
-
-@dataclass
-class CapturedData:
-    """Data captured during mission."""
-    capture_id: str
-    capture_type: CaptureType
-    timestamp: datetime
-    position: GeoPoint
-    file_path: Optional[str] = None
-    file_size_bytes: int = 0
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-    def to_dict(self) -> Dict[str, Any]:
-        return {
-            "capture_id": self.capture_id,
-            "capture_type": self.capture_type.value,
-            "timestamp": self.timestamp.isoformat(),
-            "position": self.position.to_dict(),
-            "file_path": self.file_path,
-            "file_size_bytes": self.file_size_bytes,
-            "metadata": self.metadata
-        }
-
-
-@dataclass
-class DroneState:
-    """Current state of a drone."""
-    drone_id: str
-    status: DroneStatus
-    position: Optional[GeoPoint] = None
-    battery_percent: float = 100.0
-    signal_strength: float = 100.0
-    speed: float = 0.0
-    heading: float = 0.0
-    gimbal_pitch: float = 0.0
-    gps_satellites: int = 0
-    wind_speed: float = 0.0
-    temperature: float = 20.0
-    last_update: datetime = field(default_factory=datetime.now)
-
-    def to_dict(self) -> Dict[str, Any]:
-        return {
-            "drone_id": self.drone_id,
-            "status": self.status.value,
-            "position": self.position.to_dict() if self.position else None,
-            "battery_percent": self.battery_percent,
-            "signal_strength": self.signal_strength,
-            "speed": self.speed,
-            "heading": self.heading,
-            "gimbal_pitch": self.gimbal_pitch,
-            "gps_satellites": self.gps_satellites,
-            "wind_speed": self.wind_speed,
-            "temperature": self.temperature,
-            "last_update": self.last_update.isoformat()
-        }
-
-
-@dataclass
-class Mission:
-    """Drone inspection mission."""
-    mission_id: str
-    mission_type: MissionType
-    station_id: str
-    status: MissionStatus
-    created_at: datetime
-    scheduled_at: Optional[datetime] = None
-    started_at: Optional[datetime] = None
-    completed_at: Optional[datetime] = None
-    drone_id: Optional[str] = None
-    flight_path: Optional[FlightPath] = None
-    captured_data: List[CapturedData] = field(default_factory=list)
-    home_position: Optional[GeoPoint] = None
-    current_waypoint_index: int = 0
-    progress_percent: float = 0.0
-    notes: str = ""
-    metadata: Dict[str, Any] = field(default_factory=dict)
-
-    def to_dict(self) -> Dict[str, Any]:
-        return {
-            "mission_id": self.mission_id,
-            "mission_type": self.mission_type.value,
-            "station_id": self.station_id,
-            "status": self.status.value,
-            "created_at": self.created_at.isoformat(),
-            "scheduled_at": self.scheduled_at.isoformat() if self.scheduled_at else None,
-            "started_at": self.started_at.isoformat() if self.started_at else None,
-            "completed_at": self.completed_at.isoformat() if self.completed_at else None,
-            "drone_id": self.drone_id,
-            "flight_path": self.flight_path.to_dict() if self.flight_path else None,
-            "captured_data_count": len(self.captured_data),
-            "current_waypoint_index": self.current_waypoint_index,
-            "progress_percent": self.progress_percent,
-            "notes": self.notes
-        }
-
-
-class FlightPathPlanner:
-    """Plans optimal flight paths for inspections."""
-
-    # Standard inspection patterns
-    PATTERNS = {
-        MissionType.SITE_INSPECTION: {
-            "altitude": 30,
-            "orbit_radius": 50,
-            "capture_interval": 5,  # seconds
-            "gimbal_pitch": -45
-        },
-        MissionType.TOWER_INSPECTION: {
-            "altitude_start": 5,
-            "altitude_end": 50,
-            "altitude_step": 5,
-            "orbit_radius": 15,
-            "capture_interval": 3,
-            "gimbal_pitch": 0
-        },
-        MissionType.ANTENNA_INSPECTION: {
-            "altitude": 35,
-            "orbit_radius": 10,
-            "hover_time": 5,
-            "gimbal_pitch": -30
-        },
-        MissionType.THERMAL_SCAN: {
-            "altitude": 20,
-            "grid_spacing": 10,
-            "overlap_percent": 30,
-            "gimbal_pitch": -90
-        }
-    }
-
-    def plan_mission(self,
-                     mission_type: MissionType,
-                     station_location: GeoPoint,
-                     custom_params: Optional[Dict] = None) -> FlightPath:
-        """Generate a flight path for a mission."""
-        path_id = str(uuid.uuid4())
-        params = self.PATTERNS.get(mission_type, self.PATTERNS[MissionType.SITE_INSPECTION])
-
-        if custom_params:
-            params.update(custom_params)
-
-        waypoints = []
-
-        if mission_type == MissionType.TOWER_INSPECTION:
-            waypoints = self._plan_tower_spiral(station_location, params)
-        elif mission_type == MissionType.THERMAL_SCAN:
-            waypoints = self._plan_grid_pattern(station_location, params)
-        else:
-            waypoints = self._plan_orbit_pattern(station_location, params)
-
-        # Calculate totals
-        total_distance = self._calculate_total_distance(waypoints)
-        estimated_duration = self._estimate_duration(waypoints, total_distance)
-
-        return FlightPath(
-            path_id=path_id,
-            waypoints=waypoints,
-            total_distance=total_distance,
-            estimated_duration=estimated_duration
-        )
-
-    def _plan_orbit_pattern(self, center: GeoPoint, params: Dict) -> List[Waypoint]:
-        """Plan an orbital inspection pattern."""
-        waypoints = []
-        altitude = params.get("altitude", 30)
-        radius = params.get("orbit_radius", 50)
-        num_points = 8
-
-        for i in range(num_points + 1):  # +1 to close the orbit
-            angle = (2 * np.pi * i) / num_points
-
-            # Calculate offset in meters, convert to lat/lon
-            dx = radius * np.cos(angle)
-            dy = radius * np.sin(angle)
-
-            # Approximate conversion (valid for small distances)
-            lat_offset = dy / 111000
-            lon_offset = dx / (111000 * np.cos(np.radians(center.latitude)))
-
-            position = GeoPoint(
-                latitude=center.latitude + lat_offset,
-                longitude=center.longitude + lon_offset,
-                altitude=altitude,
-                heading=(angle * 180 / np.pi + 90) % 360  # Face center
-            )
-
-            waypoint = Waypoint(
-                waypoint_id=f"wp_{i}",
-                position=position,
-                action="capture" if i < num_points else "flyover",
-                capture_type=CaptureType.PHOTO if i < num_points else None,
-                gimbal_pitch=params.get("gimbal_pitch", -45)
-            )
-            waypoints.append(waypoint)
-
-        return waypoints
-
-    def _plan_tower_spiral(self, center: GeoPoint, params: Dict) -> List[Waypoint]:
-        """Plan a spiral pattern for tower inspection."""
-        waypoints = []
-        alt_start = params.get("altitude_start", 5)
-        alt_end = params.get("altitude_end", 50)
-        alt_step = params.get("altitude_step", 5)
-        radius = params.get("orbit_radius", 15)
-
-        altitude = alt_start
-        wp_index = 0
-
-        while altitude <= alt_end:
-            num_points = 4  # Points per level
-            for i in range(num_points):
-                angle = (2 * np.pi * i) / num_points
-
-                dx = radius * np.cos(angle)
-                dy = radius * np.sin(angle)
-
-                lat_offset = dy / 111000
-                lon_offset = dx / (111000 * np.cos(np.radians(center.latitude)))
-
-                position = GeoPoint(
-                    latitude=center.latitude + lat_offset,
-                    longitude=center.longitude + lon_offset,
-                    altitude=altitude,
-                    heading=(angle * 180 / np.pi + 90) % 360
-                )
-
-                waypoint = Waypoint(
-                    waypoint_id=f"wp_{wp_index}",
-                    position=position,
-                    action="capture",
-                    capture_type=CaptureType.PHOTO,
-                    gimbal_pitch=params.get("gimbal_pitch", 0),
-                    hover_time=2.0
-                )
-                waypoints.append(waypoint)
-                wp_index += 1
-
-            altitude += alt_step
-
-        return waypoints
-
-    def _plan_grid_pattern(self, center: GeoPoint, params: Dict) -> List[Waypoint]:
-        """Plan a grid pattern for area coverage."""
-        waypoints = []
-        altitude = params.get("altitude", 20)
-        spacing = params.get("grid_spacing", 10)
-        grid_size = 5  # 5x5 grid
-
-        wp_index = 0
-        for row in range(grid_size):
-            for col in range(grid_size):
-                # Serpentine pattern
-                if row % 2 == 0:
-                    actual_col = col
-                else:
-                    actual_col = grid_size - 1 - col
-
-                dx = (actual_col - grid_size // 2) * spacing
-                dy = (row - grid_size // 2) * spacing
-
-                lat_offset = dy / 111000
-                lon_offset = dx / (111000 * np.cos(np.radians(center.latitude)))
-
-                position = GeoPoint(
-                    latitude=center.latitude + lat_offset,
-                    longitude=center.longitude + lon_offset,
-                    altitude=altitude
-                )
-
-                waypoint = Waypoint(
-                    waypoint_id=f"wp_{wp_index}",
-                    position=position,
-                    action="capture",
-                    capture_type=CaptureType.THERMAL,
-                    gimbal_pitch=params.get("gimbal_pitch", -90)
-                )
-                waypoints.append(waypoint)
-                wp_index += 1
-
-        return waypoints
-
-    def _calculate_total_distance(self, waypoints: List[Waypoint]) -> float:
-        """Calculate total flight distance."""
-        total = 0.0
-        for i in range(1, len(waypoints)):
-            total += waypoints[i-1].position.distance_to(waypoints[i].position)
-        return total
-
-    def _estimate_duration(self, waypoints: List[Waypoint], total_distance: float) -> float:
-        """Estimate mission duration in seconds."""
-        avg_speed = 5.0  # m/s
-        flight_time = total_distance / avg_speed
-
-        hover_time = sum(w.hover_time for w in waypoints)
-        capture_time = sum(3.0 for w in waypoints if w.capture_type)  # 3s per capture
-
-        return flight_time + hover_time + capture_time
+# Data models extracted to service/drone_models.py; re-exported for importers.
+from .drone_models import (  # noqa: F401
+    DroneStatus, MissionType, MissionStatus, CaptureType,
+    GeoPoint, Waypoint, FlightPath, CapturedData, DroneState, Mission,
+)
+
+
+# Flight-path planner extracted to service/flight_path_planner.py; re-exported.
+from .flight_path_planner import FlightPathPlanner  # noqa: F401
 
 
 class DroneController:
